--- practice-api/src/main/java/META-INF/additional-spring-configuration-metadata.json
@@ -1,2 +1,0 @@
-{"properties": [
-]}
