+++ practice-api/src/main/java/hjho/prj/prct/biz/mgr/mapper/MgrEmpMgMapper.java
@@ -1,0 +1,15 @@
+package hjho.prj.prct.biz.mgr.mapper;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Mapper;
+
+import hjho.prj.prct.biz.mgr.model.MgrEmpMgPagingPVO;
+import hjho.prj.prct.biz.mgr.model.MgrEmpMgPagingRVO;
+
+@Mapper
+public interface MgrEmpMgMapper {
+
+	List<MgrEmpMgPagingRVO> getMgrEmp(MgrEmpMgPagingPVO mgrDeEmpMgPagingPVO);
+
+}
