+++ practice-api/src/main/java/hjho/prj/prct/biz/place/model/LocMgPagingPVO.java
@@ -1,0 +1,15 @@
+package hjho.prj.prct.biz.place.model;
+
+import hjho.prj.prct.common.clazz.PagingModel;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@EqualsAndHashCode(callSuper=true)
+@ToString(callSuper=true)
+public class LocMgPagingPVO extends PagingModel {
+
+	private static final long serialVersionUID = 1L;
+
+}
