+++ practice-api/src/main/java/hjho/prj/prct/biz/place/model/LocMgVO.java
@@ -1,0 +1,15 @@
+package hjho.prj.prct.biz.place.model;
+
+import hjho.prj.prct.biz.table.Locations;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@EqualsAndHashCode(callSuper=true)
+@ToString(callSuper=true)
+public class LocMgVO extends Locations {
+
+	private static final long serialVersionUID = 1L;
+
+}
