--- practice-api/src/main/java/hjho/prj/prct/biz/system/model/MgrDeptMgVO.java
@@ -1,15 +1,0 @@
-package hjho.prj.prct.biz.system.model;
-
-import hjho.prj.prct.biz.table.Departments;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-
-@Data
-@EqualsAndHashCode(callSuper=true)
-@ToString(callSuper=true)
-public class MgrDeptMgVO extends Departments {
-
-	private static final long serialVersionUID = 1L;
-
-}
