--- practice-api/src/main/java/hjho/prj/prct/biz/system/service/MgrEmpMgService.java
@@ -1,22 +1,0 @@
-package hjho.prj.prct.biz.system.service;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import hjho.prj.prct.biz.system.mapper.MgrEmpMgMapper;
-import hjho.prj.prct.biz.system.model.MgrEmpMgPagingPVO;
-import hjho.prj.prct.biz.system.model.MgrEmpMgPagingRVO;
-
-@Service
-public class MgrEmpMgService {
-	
-	@Autowired 
-	private MgrEmpMgMapper mgrEmpMgMapper;
-	
-	public List<MgrEmpMgPagingRVO> getMgrEmp(MgrEmpMgPagingPVO mgrDeEmpMgPagingPVO) {
-		return mgrEmpMgMapper.getMgrEmp(mgrDeEmpMgPagingPVO);
-	} 
-	
-}
