+++ practice-api/src/main/resources/mapper/mgr/MgrDeptMgMapper.xml
@@ -1,0 +1,88 @@
+<?xml version="1.0" encoding="UTF-8" ?>  
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="hjho.prj.prct.biz.mgr.mapper.MgrDeptMgMapper">
+	
+    <sql id="order-mgr-dept">
+        ORDER BY
+        <choose>
+            <when test='sortCol == 0'>department_id</when>
+            <when test='sortCol == 1'>department_id</when>
+            <when test='sortCol == 2'>department_name</when>
+            <when test='sortCol == 3'>location_id</when>
+            <when test='sortCol == 4'>location_name</when>
+            <when test='sortCol == 5'>manager_id</when>
+            <when test='sortCol == 6'>manager_name</when>
+            <otherwise>department_id</otherwise>
+        </choose>
+        <if test='sortType == "asc"'> asc</if>
+        <if test='sortType == "desc"'> desc</if>
+        , department_id asc
+    </sql>
+	
+    <select id="getMgrDept" parameterType="hjho.prj.prct.biz.mgr.model.MgrDeptMgPagingPVO" resultType="hjho.prj.prct.biz.mgr.model.MgrDeptMgPagingRVO">
+        /* 관리자 부서 관리 > 조회 */
+        <include refid="common.sql.page-head"/>
+        SELECT department_id 
+             , department_name 
+             , manager_id 
+             , (SELECT emp.first_name||' '||emp.last_name 
+                  FROM hr.employees emp 
+                 WHERE emp.employee_id = dept.manager_id) AS manager_name
+             , location_id 
+             , (SELECT loc.city 
+                  FROM hr.locations loc 
+                 WHERE loc.location_id = dept.location_id) AS location_name 
+          FROM hr.departments dept
+        <where>
+          <if test='departmentName != null and departmentName != ""'>
+               department_name LIKE '%'||#{departmentName, jdbcType=VARCHAR}||'%'
+          </if>
+          <if test='locationId != null and locationId != ""'>
+               location_id = #{locationId, jdbcType=NUMERIC}
+          </if>
+        </where>
+        <include refid="hjho.prj.prct.biz.mgr.mapper.MgrDeptMgMapper.order-mgr-dept"/>
+        <include refid="common.sql.page-foot"/>
+    </select>
+    
+    <insert id="postMgrDept" parameterType="hjho.prj.prct.biz.mgr.model.MgrDeptMgVO">
+        /* 관리자 부서 관리 > 등록 */
+        INSERT INTO hr.departments (
+                    department_id
+                  , department_name
+                  , manager_id
+                  , location_id
+         ) VALUES ( 
+                    hr.departments_seq.NEXTVAL
+                  , #{departmentName, jdbcType=VARCHAR}
+                  , #{managerId     , jdbcType=NUMERIC}
+                  , #{locationId    , jdbcType=NUMERIC}
+         )
+                
+    </insert>
+    
+    <update id="putMgrDept" parameterType="hjho.prj.prct.biz.mgr.model.MgrDeptMgVO">
+        /* 관리자 부서 관리 > 수정 */
+        UPDATE hr.departments
+           SET department_name = #{departmentName, jdbcType=VARCHAR}
+             , manager_id      = #{managerId     , jdbcType=NUMERIC}
+             , location_id     = #{locationId    , jdbcType=NUMERIC}
+         WHERE department_id   = #{departmentId  , jdbcType=NUMERIC}
+    </update>
+    
+    <delete id="deleteMgrDept" parameterType="hjho.prj.prct.biz.mgr.model.MgrDeptMgVO">
+        /* 관리자 부서 관리 > 삭제 */
+        DELETE 
+          FROM hr.departments
+         WHERE department_id = #{departmentId, jdbcType=NUMERIC}
+    </delete>
+    
+    <select id="pkCheck" parameterType="hjho.prj.prct.biz.mgr.model.MgrDeptMgVO" resultType="int">
+        /* 관리자 부서 관리 > PK */
+        SELECT COUNT(*) AS cnt
+          FROM hr.departments
+         WHERE department_id = #{departmentId, jdbcType=NUMERIC} 
+    </select>
+    
+</mapper>
