--- practice-api/src/main/resources/mapper/system/MgrDeptMgMapper.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>  
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="hjho.prj.prct.biz.system.mapper.MgrDeptMgMapper">
-	
-    <sql id="order-mgr-dept">
-        ORDER BY
-        <choose>
-            <when test='sortCol == 0'>department_id</when>
-            <when test='sortCol == 1'>department_id</when>
-            <when test='sortCol == 2'>department_name</when>
-            <when test='sortCol == 3'>location_id</when>
-            <when test='sortCol == 4'>location_name</when>
-            <when test='sortCol == 5'>manager_id</when>
-            <when test='sortCol == 6'>manager_name</when>
-            <otherwise>department_id</otherwise>
-        </choose>
-        <if test='sortType == "asc"'> asc</if>
-        <if test='sortType == "desc"'> desc</if>
-        , department_id asc
-    </sql>
-	
-    <select id = "getMgrDept" 
-            parameterType = "hjho.prj.prct.biz.system.model.MgrDeptMgPagingPVO" 
-            resultType    = "hjho.prj.prct.biz.system.model.MgrDeptMgPagingRVO">
-        /* MgrDeptMgMapper.getMgrDept */
-    <include refid="common.sql.page-head"/>
-        SELECT department_id 
-             , department_name 
-             , manager_id 
-             , (SELECT emp.first_name||' '||emp.last_name 
-                  FROM hr.employees emp 
-                 WHERE emp.employee_id = dept.manager_id) AS manager_name
-             , location_id 
-             , (SELECT loc.city 
-                  FROM hr.locations loc 
-                 WHERE loc.location_id = dept.location_id) AS location_name 
-          FROM hr.departments dept
-      <where>
-          <if test='departmentName != null and departmentName != ""'>
-               department_name LIKE '%'||#{departmentName, jdbcType=VARCHAR}||'%'
-          </if>
-          <if test='locationId != null and locationId != ""'>
-               location_id = #{locationId, jdbcType=NUMERIC}
-          </if>
-      </where>
-    <include refid="hjho.prj.prct.biz.system.mapper.MgrDeptMgMapper.order-mgr-dept"/>
-    <include refid="common.sql.page-foot"/>
-    </select>
-    
-</mapper>
