--- practice-api/src/main/resources/mapper/system/MgrEmpMgMapper.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>  
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="hjho.prj.prct.biz.system.mapper.MgrEmpMgMapper">
-	
-    <sql id="order-mgr-emp">
-        ORDER BY
-        <choose>
-            <when test='sortCol == 0'>department_id</when>
-            <when test='sortCol == 1'>department_id</when>
-            <when test='sortCol == 2'>department_name</when>
-            <when test='sortCol == 3'>location_id</when>
-            <when test='sortCol == 4'>location_name</when>
-            <when test='sortCol == 5'>manager_id</when>
-            <when test='sortCol == 6'>manager_name</when>
-            <otherwise>department_id</otherwise>
-        </choose>
-        <if test='sortType == "asc"'> asc</if>
-        <if test='sortType == "desc"'> desc</if>
-        , department_id asc
-    </sql>
-	
-    <select id = "getMgrEmp" 
-            parameterType = "hjho.prj.prct.biz.system.model.MgrEmpMgPagingPVO" 
-            resultType    = "hjho.prj.prct.biz.system.model.MgrEmpMgPagingRVO">
-        /* MgrEmpMgMapper.getMgrEmp */
-    <include refid="common.sql.page-head"/>
-        SELECT employee_id 
-             , first_name 
-             , last_name 
-             , email 
-             , phone_number 
-             , hire_date 
-             , job_id
-             , salary
-             , commission_pct
-             , manager_id
-             , department_id
-             , (SELECT department_name 
-                  FROM hr.departments dept 
-                 WHERE dept.department_id = emp.department_id) AS department_name
-          FROM hr.employees emp
-   <!--  <include refid="hjho.prj.prct.biz.system.mapper.MgrDeptMgMapper.order-mgr-dept"/>  -->
-    <include refid="common.sql.page-foot"/>
-    </select>
-    
-</mapper>
