+++ practice-ui/src/main/java/hjho/prj/prct/biz/mgr/model/MgrDeptMgPagingVO.java
@@ -1,0 +1,19 @@
+package hjho.prj.prct.biz.mgr.model;
+
+import hjho.prj.prct.common.clazz.PagingModel;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper=true)
+@EqualsAndHashCode(callSuper=true)
+public class MgrDeptMgPagingVO extends PagingModel {
+
+	private static final long serialVersionUID = 1L;
+
+	private String departmentName;
+	
+	private Long   locationId;
+	
+}
