+++ practice-ui/src/main/java/hjho/prj/prct/biz/mgr/model/MgrDeptMgVO.java
@@ -1,0 +1,23 @@
+package hjho.prj.prct.biz.mgr.model;
+
+import hjho.prj.prct.common.clazz.CommonModel;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper=true)
+@EqualsAndHashCode(callSuper=true)
+public class MgrDeptMgVO extends CommonModel {
+
+	private static final long serialVersionUID = 1L;
+
+	private Long   departmentId;
+	
+	private String departmentName;
+	
+	private Long   managerId;
+			
+	private Long   locationId;
+	
+}
