+++ practice-ui/src/main/java/hjho/prj/prct/biz/mgr/model/MgrEmpMgPagingVO.java
@@ -1,0 +1,15 @@
+package hjho.prj.prct.biz.mgr.model;
+
+import hjho.prj.prct.common.clazz.PagingModel;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper=true)
+@EqualsAndHashCode(callSuper=true)
+public class MgrEmpMgPagingVO extends PagingModel {
+
+	private static final long serialVersionUID = 1L;
+
+}
