+++ practice-ui/src/main/java/hjho/prj/prct/biz/place/model/LocMgPagingVO.java
@@ -1,0 +1,15 @@
+package hjho.prj.prct.biz.place.model;
+
+import hjho.prj.prct.common.clazz.PagingModel;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+
+@Data
+@ToString(callSuper=true)
+@EqualsAndHashCode(callSuper=true)
+public class LocMgPagingVO extends PagingModel {
+
+	private static final long serialVersionUID = 1L;
+
+}
