--- practice-ui/src/main/java/hjho/prj/prct/biz/system/model/MgrDeptMgVO.java
@@ -1,21 +1,0 @@
-package hjho.prj.prct.biz.system.model;
-
-import hjho.prj.prct.common.clazz.CommonModel;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-@Data
-@EqualsAndHashCode(callSuper=true)
-public class MgrDeptMgVO extends CommonModel {
-
-	private static final long serialVersionUID = 1L;
-
-	private long   departmentId;
-	
-	private String departmentName;
-	
-	private long   managerId;
-			
-	private long   locationId;
-	
-}
