+++ practice-ui/src/main/resources/static/js/common/hjho-datatables.js
@@ -2,7 +2,7 @@
  * 공통 Java Script DataTable function 
  */
 var DataTableUtils = function() {
-	var lengthMenu = [[10, 20, 40, -1], [10, 20, 40, 'All']];
+	var lengthMenu = [[10, 20, 40, 80], [10, 20, 40, '80']];
 	var language = {
 		    "decimal":        "",
 		    "emptyTable":     "검색 결과가 없습니다.",
@@ -41,10 +41,39 @@ var DataTableUtils = function() {
 			    "language" : language,
 			    "autoWidth": false,
 			    "select": "single",
+			    "searching" : false,
 			    "data": []
 			});
 		},
 		get : function(data) {
+			
+			// 포멧별 컬럼 렌더링
+			$.each(data.columns, function(index, item) {
+				// column format type
+				var type = item.format;
+				
+				if(StringUtils.isNotEmpty(type)) {
+					switch(type) {
+					case "dtm"  : // 일시 
+						item["render"] = function(data, type, row) {
+							return DateUtils.dtmFormat(data);
+						}
+						break;
+					case "date" : // 일자
+						item["render"] = function(data, type, row) {
+							return DateUtils.dateFormat(data);
+						}
+						break;
+					case "time" : // 시간
+						item["render"] = function(data, type, row) {
+							return DateUtils.timeFormat(data);
+						}
+						break;
+					}
+				}
+			});
+			
+			// 데이터 테이블 RUN
 			$(data.tableId).dataTable( {
 				"destroy": true,
 				"serverSide": true,
@@ -54,6 +83,7 @@ var DataTableUtils = function() {
 				"ordering": true,
 				"autoWidth": false,
 				"select": "single",
+				"searching" : false,
 				"lengthMenu": lengthMenu,
 				"language" : language,
 				"ajax" : {
