+++ practice-ui/src/main/resources/templates/mainPage.html
@@ -5,6 +5,7 @@
 	<link rel="shortcut icon" href="#">
 	<!-- jQuery -->
 	<script th:src="@{/js/jquery/jquery-3.6.0.min.js}" type="text/javascript"></script>
+	
 	<!-- Bootstrap 4.0.0 -->
 	<link th:href="@{/bootstrap-4.0.0/css/bootstrap.css}" rel="stylesheet" type="text/css" />
 	<link th:href="@{/bootstrap-4.0.0/css/bootstrap-grid.css}" rel="stylesheet" type="text/css" />
@@ -28,6 +29,9 @@
 	<!-- 메인 페이지 -->
 	<link th:href="@{/bootstrap-4.0.0/css/main.css}" rel="stylesheet" type="text/css" />
 	<script th:src="@{/bootstrap-4.0.0/js/feather.js}" type="text/javascript"></script>
+	
+	<!-- PRACTICE -->
+	<link th:href="@{/css/practice.css}" rel="stylesheet" type="text/css" />
 	
 <title>프로젝트 연습(main)</title>
 </head>
@@ -55,7 +59,7 @@
       		
         	<div th:replace="/fragment/menuCtnt.html :: menu">
         	</div>
-        	<div role="main" class="col-md-9 ml-sm-auto col-lg-10 pt-3 px-4">
+        	<div role="main" class="col-md-9 ml-sm-auto col-lg-10 pt-3 px-4 main-page">
 	        <!-- 콘텐츠 시작 -->
 	        	<div th:replace="${viewPath} :: ${viewName}"></div>
 	        <!-- 콘텐츠 끝 -->
