+++ practice-ui/src/main/resources/templates/mgr/employeeModal.html
@@ -1,0 +1,92 @@
+<!--------------------- START EMP MODAL START ----------------------------->
+<div class="modal fade" id="modalEmployee" tabindex="-1" role="dialog" aria-hidden="true">
+ 	<div class="modal-dialog modal-lg modal-dialog-centered" role="document">
+    	<div class="modal-content">
+      		<div class="modal-header">
+        		<h5 class="modal-title">사원 선택</h5>
+        		<button type="button" class="close" data-dismiss="modal" aria-label="Close">
+          			<span aria-hidden="true">&times;</span>
+        		</button>
+      		</div>
+      		<div class="modal-body">
+      			<div class="card">
+				  	<h5 class="card-header">사원 목록</h5>
+				  	<div class="card-body">
+				  		<table class="table table-hover table-bordered" id="tableEmployee">
+							<thead class="table-light">
+								<tr>
+									<th class="text-center">No</th>
+									<th class="text-center">부서명</th>
+									<th class="text-center">사원번호</th>
+									<th class="text-center">사원명</th>
+									<th class="text-center">고용일시</th>
+									<th class="text-center">직책</th>
+								</tr>
+							</thead>
+						</table>
+				  	</div>
+				</div>
+      		</div>
+	      	<div class="modal-footer">
+	        	<button type="button" class="btn btn-secondary" data-dismiss="modal">닫기</button>
+	        	<button type="button" class="btn btn-primary" onclick="fnChocEmp()">선택</button>
+	      	</div>
+    	</div>
+  	</div>
+</div>
+<script>
+	$(document).ready(function() {
+		$('#tableEmployee').on('click', 'tbody tr', function() {
+			var data = $('#tableEmployee').dataTable().fnGetData(this);
+			if(StringUtils.isEmpty(data)) {
+				return;
+			}
+			
+			if(DataTableUtils.isSeleted('#tableEmployee', this)) {
+				return;
+			}
+		});
+	});
+	// 사원 조회
+	function fnGetMgrEmp() {
+		// alert("사원 검색");
+		DataTableUtils.get({
+			tableId : "#tableEmployee",
+			url : "/mgr/dept/emp",
+			columns: [
+	            { "data": "rowNo"       , "class": "text-center"},
+	            { "data": "departmentName", "class": "text-center"},
+	            { "data": "employeeId"  , "class": "text-center"},
+	            { 
+	            	"data": "firstName"	, "class": "text-left"  ,
+	            	"render": function(data, type, row) {
+	            		return data + " " + row.lastName;
+	            	}
+	            },
+	            { "data": "hireDate"    , "class": "text-center", "format": "dtm"},
+	            { "data": "jobId"  		, "class": "text-left"  },
+	        ]
+		});
+	}
+	// 사원 선택
+	function fnChocEmp() {
+		var data = $('#tableEmployee').dataTable().fnGetData('#tableEmployee .selected');
+		if(StringUtils.isEmpty(data)) {
+			return alert("사원을 선택해주세요.");;
+		}
+		
+		var empName = data.firstName + " " + data.lastName;
+		var empId   = data.employeeId;
+		
+		var message =  empName + "(" +empId+ ")";
+		    message += "사원을 선택하시겠습니까?";
+		if(confirm(message)) {
+			// 관리자ID/명
+			// alert("사원 번호 : "+ empId + ", 사원 명 : " + empName);
+			$('#formMgrDept').find('#managerId').val(empId);
+			$('#formMgrDept').find('#managerName').val(empName);
+			$('#modalEmployee').modal('hide');
+		}
+	}
+</script>
+<!--------------------- END EMP MODAL END ----------------------------->
